Remove restricted_inheritance option from Chex dataclasses
This removes the optional check on dataclasses that they only derive
from object or dataclass. This is a hold-over from a previous
implementation of mappable dataclasses which is now unnecessary.

The check was added because the dataclass's `__bases__` attribute was
replaced with `abc.Mappable`, which has the effect of removing
behaviour coming from a superclass. This was to avoid adding both
`object` and `abc.Mappable` to `__bases__`, which causes an MRO
error. Later on, the `abc.Mappable` behaviour was inserted into
`__bases__` (and `object` removed if present), which makes the check
unnecessary.

A warning message is logged if the dataclass overrides `__contains__`,
`__eq__`, or `__ne__` since these implementations will be used over
the mixin implementations from `Mapping`, which might confuse the
user.

This change was prompted by a request to support mappable dataclasses
deriving from Generic, so this commit contains a unit test asserting
that this is now possible.

PiperOrigin-RevId: 393112153